Implement method to copy multiple modules
The current Beaker DSL has command to install and copy modules to
one or more target host(s), but does not feature installing multiple
modules at once while this could be useful for example to use the
same modulesets for rspec-puppet tests and acceptance tests.

This commit adds a new method copy_modules_to to the DSL,
which takes a source_dir and a list of module names and
invokes copy_module_to for each module name.